# api.gen.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package generated

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
)

// CreateOrderRequest defines model for CreateOrderRequest.
type CreateOrderRequest struct {
	CardToken    string      `json:"card_token"`
	CustomerId   string      `json:"customer_id"`
	Items        []OrderItem `json:"items"`
	RestaurantId string      `json:"restaurant_id"`
}

// OrderCreatedResponse defines model for OrderCreatedResponse.
type OrderCreatedResponse struct {
	Id     *string `json:"id,omitempty"`
	Status *string `json:"status,omitempty"`
}

// OrderItem defines model for OrderItem.
type OrderItem struct {
	Name      string  `json:"name"`
	Price     float32 `json:"price"`
	ProductId string  `json:"product_id"`
	Quantity  int     `json:"quantity"`
}

// OrderResponse defines model for OrderResponse.
type OrderResponse struct {
	CreatedAt    *time.Time `json:"created_at,omitempty"`
	CustomerId   *string    `json:"customer_id,omitempty"`
	Id           *string    `json:"id,omitempty"`
	RestaurantId *string    `json:"restaurant_id,omitempty"`
	Status       *string    `json:"status,omitempty"`
	TotalAmount  *float32   `json:"total_amount,omitempty"`
}

// CreateOrderJSONRequestBody defines body for CreateOrder for application/json ContentType.
type CreateOrderJSONRequestBody = CreateOrderRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Create a new order
	// (POST /orders)
	CreateOrder(ctx echo.Context) error
	// Get order details and status
	// (GET /orders/{id})
	GetOrder(ctx echo.Context, id string) error
	// Cancel an order
	// (POST /orders/{id}/cancel)
	CancelOrder(ctx echo.Context, id string) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// CreateOrder converts echo context to params.
func (w *ServerInterfaceWrapper) CreateOrder(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateOrder(ctx)
	return err
}

// GetOrder converts echo context to params.
func (w *ServerInterfaceWrapper) GetOrder(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetOrder(ctx, id)
	return err
}

// CancelOrder converts echo context to params.
func (w *ServerInterfaceWrapper) CancelOrder(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CancelOrder(ctx, id)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/orders", wrapper.CreateOrder)
	router.GET(baseURL+"/orders/:id", wrapper.GetOrder)
	router.POST(baseURL+"/orders/:id/cancel", wrapper.CancelOrder)

}

type CreateOrderRequestObject struct {
	Body *CreateOrderJSONRequestBody
}

type CreateOrderResponseObject interface {
	VisitCreateOrderResponse(w http.ResponseWriter) error
}

type CreateOrder202JSONResponse OrderCreatedResponse

func (response CreateOrder202JSONResponse) VisitCreateOrderResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(202)

	return json.NewEncoder(w).Encode(response)
}

type CreateOrder400Response struct {
}

func (response CreateOrder400Response) VisitCreateOrderResponse(w http.ResponseWriter) error {
	w.WriteHeader(400)
	return nil
}

type GetOrderRequestObject struct {
	Id string `json:"id"`
}

type GetOrderResponseObject interface {
	VisitGetOrderResponse(w http.ResponseWriter) error
}

type GetOrder200JSONResponse OrderResponse

func (response GetOrder200JSONResponse) VisitGetOrderResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetOrder404Response struct {
}

func (response GetOrder404Response) VisitGetOrderResponse(w http.ResponseWriter) error {
	w.WriteHeader(404)
	return nil
}

type CancelOrderRequestObject struct {
	Id string `json:"id"`
}

type CancelOrderResponseObject interface {
	VisitCancelOrderResponse(w http.ResponseWriter) error
}

type CancelOrder204Response struct {
}

func (response CancelOrder204Response) VisitCancelOrderResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type CancelOrder400Response struct {
}

func (response CancelOrder400Response) VisitCancelOrderResponse(w http.ResponseWriter) error {
	w.WriteHeader(400)
	return nil
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Create a new order
	// (POST /orders)
	CreateOrder(ctx context.Context, request CreateOrderRequestObject) (CreateOrderResponseObject, error)
	// Get order details and status
	// (GET /orders/{id})
	GetOrder(ctx context.Context, request GetOrderRequestObject) (GetOrderResponseObject, error)
	// Cancel an order
	// (POST /orders/{id}/cancel)
	CancelOrder(ctx context.Context, request CancelOrderRequestObject) (CancelOrderResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// CreateOrder operation middleware
func (sh *strictHandler) CreateOrder(ctx echo.Context) error {
	var request CreateOrderRequestObject

	var body CreateOrderJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.CreateOrder(ctx.Request().Context(), request.(CreateOrderRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateOrder")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(CreateOrderResponseObject); ok {
		return validResponse.VisitCreateOrderResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetOrder operation middleware
func (sh *strictHandler) GetOrder(ctx echo.Context, id string) error {
	var request GetOrderRequestObject

	request.Id = id

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetOrder(ctx.Request().Context(), request.(GetOrderRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetOrder")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetOrderResponseObject); ok {
		return validResponse.VisitGetOrderResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// CancelOrder operation middleware
func (sh *strictHandler) CancelOrder(ctx echo.Context, id string) error {
	var request CancelOrderRequestObject

	request.Id = id

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.CancelOrder(ctx.Request().Context(), request.(CancelOrderRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CancelOrder")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(CancelOrderResponseObject); ok {
		return validResponse.VisitCancelOrderResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}