# ordering/internal/adapters/db/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.25.0

package sqlc

import (
	"database/sql"
	"encoding/json"
	"time"
)

type Order struct {
	ID              string    `json:"id"`
	CustomerID      string    `json:"customer_id"`
	RestaurantID    string    `json:"restaurant_id"`
	Status          string    `json:"status"`
	TotalAmount     int64     `json:"total_amount"`
	Currency        string    `json:"currency"`
	DeliveryAddress string    `json:"delivery_address"`
	CorrelationID   string    `json:"correlation_id"`
	Version         int32     `json:"version"`
	CreatedAt       time.Time `json:"created_at"`
	UpdatedAt       time.Time `json:"updated_at"`
}

type OrderItem struct {
	ID         string         `json:"id"`
	OrderID    string         `json:"order_id"`
	MenuItemID string         `json:"menu_item_id"`
	Name       string         `json:"name"`
	Quantity   int32          `json:"quantity"`
	UnitPrice  int64          `json:"unit_price"`
	Notes      sql.NullString `json:"notes"`
}

type OutboxMessage struct {
	ID            string          `json:"id"`
	AggregateType string          `json:"aggregate_type"`
	AggregateID   string          `json:"aggregate_id"`
	EventType     string          `json:"event_type"`
	Payload       json.RawMessage `json:"payload"`
	CorrelationID string          `json:"correlation_id"`
	CreatedAt     time.Time       `json:"created_at"`
	PublishedAt   sql.NullTime    `json:"published_at"`
}

type SagaState struct {
	OrderID     string          `json:"order_id"`
	CurrentStep string          `json:"current_step"`
	Status      string          `json:"status"`
	DataJson    json.RawMessage `json:"data_json"`
	UpdatedAt   time.Time       `json:"updated_at"`
}